refactor(wallet): move all contracts under Interface namespace (#566)
- BRC100::Interface → Interface::BRC100 (single file, 28 stubs)
- Store interface methods → Interface::Store
- BroadcastQueue interface methods → Interface::BroadcastQueue
- ProofStore → Interface::ProofStore
- UTXOPool → Interface::UTXOPool

Store and BroadcastQueue modules become pure namespace hubs for
implementations. The Interface module is the single location for
all abstract contracts.

Collapse 6 BRC100 sub-module files into one — structure for
structure's sake serves no purpose when each file is 3-5 stubs.
