+++ gem/bsv-wallet-postgres/spec/bsv/wallet_postgres/solid_queue_adapter_spec.rb
@@ -131,7 +131,7 @@ RSpec.describe 'BSV::Wallet::SolidQueueAdapter', :postgres do
   describe 'BroadcastQueue inclusion' do
     it 'includes BroadcastQueue' do
       expect(described_class).to be_nil # no-op; verify via ancestors
-      expect(BSV::Wallet::SolidQueueAdapter.ancestors).to include(BSV::Wallet::BroadcastQueue)
+      expect(BSV::Wallet::SolidQueueAdapter.ancestors).to include(BSV::Wallet::Interface::BroadcastQueue)
     end
   end
