+++ gem/bsv-wallet/lib/bsv/wallet/interface.rb
@@ -2,13 +2,16 @@
 
 module BSV
   module Wallet
-    # BRC-100 Wallet Interface
+    # Abstract interfaces (contracts) for the wallet ecosystem.
     #
-    # Composes all 28 BRC-100 wallet methods from the {BRC100::Interface} sub-modules.
-    # Include this module and override the methods your implementation supports.
-    # Unimplemented methods raise {UnsupportedActionError}.
+    # Each module under Interface defines what an implementation must do,
+    # not how. Include the relevant interface in your concrete class.
     module Interface
-      include BRC100::Interface
+      autoload :BRC100,         'bsv/wallet/interface/brc100'
+      autoload :Store,          'bsv/wallet/interface/store'
+      autoload :BroadcastQueue, 'bsv/wallet/interface/broadcast_queue'
+      autoload :ProofStore,     'bsv/wallet/interface/proof_store'
+      autoload :UTXOPool,       'bsv/wallet/interface/utxo_pool'
     end
   end
 end
