+++ gem/bsv-wallet/spec/support/shared_examples_for_storage_adapter.rb
@@ -1,11 +1,11 @@
 # frozen_string_literal: true
 
-# Shared conformance specs for BSV::Wallet::Store implementations.
+# Shared conformance specs for BSV::Wallet::Interface::Store implementations.
 #
-# Any class that includes +BSV::Wallet::Store+ should pass this
+# Any class that includes +BSV::Wallet::Interface::Store+ should pass this
 # suite. Callers provide a +store+ via +let(:store)+:
 #
-#   RSpec.describe BSV::Wallet::Store::Memory do
+#   RSpec.describe BSV::Wallet::Interface::Store::Memory do
 #     let(:store) { described_class.new }
 #     it_behaves_like 'a storage adapter'
 #   end
@@ -14,7 +14,7 @@
 # proofs, transactions, count_*, and pagination.
 RSpec.shared_examples 'a storage adapter' do
   it 'includes Store' do
-    expect(store.class.ancestors).to include(BSV::Wallet::Store)
+    expect(store.class.ancestors).to include(BSV::Wallet::Interface::Store)
   end
 
   describe 'actions' do
